renameat2: add EL7 compatibility handling
EL7 has backported renameat2(2) support, but this was done by creating a
wrapper struct around the inode_operations struct and adding the rename2
callback to the wrapping struct. The semantics are the same as the
upstream callback, so just detect and use the wrapper.

Since we only need this wrapper for rename2 (and renaming is a directory
operation) we only use the wrapper for the directory iops struct.

Signed-off-by: Pavel Snajdr <snajpa@snajpa.net>
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>