Use setproctitle to report progress of zfs send
This allows parsing of zfs send progress by checking the process
title.

Doing so requires some changes to the send code in libzfs_sendrecv.c;
primarily these changes move some of the accounting around, to allow
for the code to be verbose as normal, or set the process title. Unlike
BSD, setproctitle() isn't standard in Linux; I found a reference to it
in libbsd, and included autoconf-related changes to test for that.

Authored-by: Sean Eric Fagan <sef@FreeBSD.org>
Signed-off-by: Ryan Moeller <ryan@iXsystems.com>