fix crash in make_dataset_handle_common
Running the zettarepl integration tests on TrueNAS will produce core
dumps for the middlewared process 100% of the time.

Crash is occurring in make_dataset_handle_common because `zfs_type` is
`DMU_OST_NONE`. Unfortunately, I haven't been successful in tracking
down whether or not this is a valid `zfs_type`. However, 100% of the
crashes that I've investigated from end-users as well as the ones that
were generated by the zettarepl integration tests show that
`dds_inconsistent` is 1. This means the filesystem is in the process of
being received or some case of destruction. Investigation has shown that
other parts of the libzfs API also have similar logic.

This changes `make_dataset_handle_common` to gracefully handle the
situation when `dds_inconsistent` != 0 and simply return -1 as to not
cause an `abort()`. With this change, the zettarepl integration tests
complete successfully and no more crashes occur.