Fix bash_completion installation
bash_completion is present in TrueNAS ZFS but it was not getting
installed in to any of the packages. This commit fixes this issue
with both native Debian packages and RPM converted Debian packages.

Upstream does not have this issue, as they have updated the make
build process and it is fixed in there.

Signed-off-by: Umer Saleem <usaleem@ixsystems.com>