Add Provides field for packages in contrib/debian/control
For TrueNAS SCALE, there are packages that depend on OpenZFS
packages. Since the package names are changed in native Debian
packaging, scale-build fails to find the OpenZFS packages, as it
is looking for libnvpair3, libuutilt3 etc.

Instead of updating the Depends section in control file for each
individual package that depends on OpenZFS packages, this commit
adds Provides field for native packages. In addition to package
names, we can now look for virtual packages with names added to
Provides field.

RPM converted package names are added to Provides field, so that
scale-build can identify, look for and install package dependencies.

Signed-off-by: Umer Saleem <usaleem@ixsystems.com>