Add '-u' - nomount flag for zfs set
This commit adds '-u' flag for zfs set operation. With this flag,
mountpoint, sharenfs and sharesmb properties can be updated
without actually mounting or sharing the dataset.

Previously, if dataset was unmounted, and mountpoint property was
updated, dataset was not mounted after the update. This behavior
is changed in previous commit. We now mount the dataset whenever
mountpoint property is updated, regardless if it's mounted or not.

To provide the user with option to keep the dataset unmounted and
still update the mountpoint without mounting the dataset, '-u'
flag can be used.

While unmounted, if sharenfs or sharesmb is set to on, this leads
to mounting the dataset. If sharenfs and sharesmb properties are
set with '-u' flag while dataset is unmounted, this will not mount
or share the dataset, but it will set the property to desired
value.

When dataset is mounted, and '-u' flag is used to update any of
mountpoint, sharenfs and sharesmb property/properties, the
property itself is updated but no action is performed. For
mountpoint property, the property is updated but dataset is still
mounted at previous mountpoint. For sharenfs and sharesmb, the
property is set but dataset in not shared/unshared.

'-u' flag would only work if any of mountpoint, sharenfs or
sharesmb is found in property list.

Signed-off-by: Umer Saleem <usaleem@ixsystems.com>